Déplacement de la section Les démarches à réaliser de ServiceMobilize à ServiceSteps
